feat: Move map zoom functions to MapsModel
