+++ hackathon-for-hunger/DriverMapDetailPendingVC.swift
@@ -13,6 +13,7 @@ import CoreLocation
 class DriverMapDetailPendingVC: UIViewController, MKMapViewDelegate {
     
     var startingRegion = MapsDummyData.sharedInstance.startingRegion // used to retrieve precalculated starting region
+    var mapsModel = MapsModel.sharedInstance
     
     //var annotationsHaveBeenShown: Bool = false //to allow auto zooming to pins, but just once
     
@@ -87,38 +88,11 @@ class DriverMapDetailPendingVC: UIViewController, MKMapViewDelegate {
             return
         }
         
-        var topLeftCoord = CLLocationCoordinate2D(latitude: -90, longitude: 180)
-        var bottomRightCoord = CLLocationCoordinate2D(latitude: 90, longitude: -180)
+        let region = mapsModel.getRegionForAnnotations(mapView.annotations)
         
-        for annotation in mapView.annotations {
-            topLeftCoord.longitude = min(topLeftCoord.longitude, annotation.coordinate.longitude)
-            topLeftCoord.latitude = max(topLeftCoord.latitude, annotation.coordinate.latitude)
-            bottomRightCoord.longitude = max(bottomRightCoord.longitude, annotation.coordinate.longitude)
-            bottomRightCoord.latitude = min(bottomRightCoord.latitude, annotation.coordinate.latitude)
-        }
-        
-        var region = MKCoordinateRegion()
-        region.center.latitude = topLeftCoord.latitude - (topLeftCoord.latitude - bottomRightCoord.latitude) * 0.5
-        region.center.longitude = topLeftCoord.longitude + (bottomRightCoord.longitude - topLeftCoord.longitude) * 0.5
-        
-        // Add a little extra space on the sides
-        region.span.latitudeDelta = abs(topLeftCoord.latitude - bottomRightCoord.latitude) * 1.3
-        region.span.longitudeDelta = abs(bottomRightCoord.longitude - topLeftCoord.longitude) * 1.3
-        
-        region = mapView.regionThatFits(region)
-        //mapView.setRegion(region, animated: true)
-        mapView.setVisibleMapRect(MKMapRectForCoordinateRegion(region), edgePadding: UIEdgeInsets(top: 50, left: 50, bottom: 0, right: 50), animated: true)
+        mapView.setVisibleMapRect(mapsModel.MKMapRectForCoordinateRegion(mapView.regionThatFits(region)), edgePadding: UIEdgeInsets(top: 50, left: 50, bottom: 0, right: 50), animated: true)
     }
     
-    func MKMapRectForCoordinateRegion(region:MKCoordinateRegion) -> MKMapRect {
-        let topLeft = CLLocationCoordinate2D(latitude: region.center.latitude + (region.span.latitudeDelta/2), longitude: region.center.longitude - (region.span.longitudeDelta/2))
-        let bottomRight = CLLocationCoordinate2D(latitude: region.center.latitude - (region.span.latitudeDelta/2), longitude: region.center.longitude + (region.span.longitudeDelta/2))
-        
-        let a = MKMapPointForCoordinate(topLeft)
-        let b = MKMapPointForCoordinate(bottomRight)
-        
-        return MKMapRect(origin: MKMapPoint(x:min(a.x,b.x), y:min(a.y,b.y)), size: MKMapSize(width: abs(a.x-b.x), height: abs(a.y-b.y)))
-    }
     
     
     //MARK:- mapView
